refactor(app) : refactor search page
- split component into small component to improve reusability and readability
